Correct StringResourceId to u16 and narrow retail mappings
Win32 LoadString uses LOWORD only, so StringResourceId is now a wrapping
u16. Factor the shared RT_STRING decode, delegate UI string policy through
RetailAssetsResource, keep retail_resources for reusable semantic mappings
only, restore cohesive screen-local matches, and collapse identity retail()
ordinals in core.

+++ rust/crates/imperialism-app/src/ui/credits.rs
@@ -103,7 +103,7 @@ fn fill_credits_page(
 
 fn string_from_id(assets: &RetailUiAssets, string_id: i16) -> String {
     // `LoadUiStringResourceById` is LoadStringA(id); the flat id is the final resource id.
-    assets.string(StringResourceId::new(string_id as u32))
+    assets.string(StringResourceId::new(string_id as u16))
 }
 
 fn on_credits_activate(
